feat: clear button, filters and result count on one line.
