+++ shesha-reactjs/src/designer-components/dataTable/table/styles/styles.ts
@@ -1,4 +1,4 @@
-import { createGlobalStyle, createStyles } from "antd-style";
+import { createGlobalStyle } from "antd-style";
 import { sheshaStyles } from '@/styles';
 
 
@@ -79,88 +79,3 @@ export const GlobalTableStyles = createGlobalStyle`
     }
   }
 `;
-
-
-export const useStyles = createStyles(({ token, cx, css }) => {
-
-  const primaryColor = token.colorPrimary;
-  const secondaryColor = token.colorPrimaryBgHover;
-  const arrowLeft = "scroll-arrow-left";
-  const arrowRight = "scroll-arrow-right";
-  const tag = "tag";
-
-  const wrapper = cx("filter - wrapper", css`
-      display: flex;
-      whiteSpace: nowrap;
-      justifyContent: center;
-      `);
-
-  const scrollableTagsContainer = cx("scrollable-tags-container", css`
-      max-width: 700px;
-      margin: 6px;
-      justify-content: center;
-      overflow: hidden;
-      display: flex;
-      position: relative;
-      user-select: none;
-      ::-webkit-scrollbar {
-        display: none;
-      }
-
-      * {
-        -ms-overflow-style: none; /* for Internet Explorer, Edge */
-        scrollbar-width: none; /* for Firefox */
-      }
-
-      p {
-        white-space: nowrap;
-      }
-      .filters {
-        display: flex;
-        margin: 0 24px;
-        overflow-x: scroll;
-        scroll-behavior: smooth;
-        align-items: center;
-      }
-
-      .${arrowLeft}, .${arrowRight} {
-        color: ${token.colorPrimary};
-        cursor: pointer;
-        position: absolute;
-        height: 100%;
-        top: 0;
-        width: 24px;
-        justify-content: center;
-        display: flex;
-      }
-
-      .hidden {
-        display: none;
-      }
-
-      .${arrowLeft} {
-        left: 0;
-        border-radius: 4px 0 0 4px;
-      }
-
-       .${arrowRight} {
-        right: 0;
-        border-radius: 0 4px 4px 0;
-      }
-      
-      .${tag} {
-        color: ${token.colorPrimary};
-        marginBottom: .32em;
-        vertical-align: middle;
-    `);
-
-  return {
-    secondaryColor,
-    primaryColor,
-    arrowLeft,
-    arrowRight,
-    tag,
-    wrapper,
-    scrollableTagsContainer
-  };
-});
